requirements.txt
python-telegram-bot>=21.0
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
ta>=0.11.0
candlestick==0.0.8
python-dotenv>=1.0.0
schedule>=1.2.1
